push: reorganize code, light refactoring
