+++ push/header.go
@@ -1,0 +1,42 @@
+package push
+
+import (
+	"net/http"
+	"strconv"
+	"time"
+)
+
+// Headers sent with a push to control the notification (optional)
+type Headers struct {
+	// ID for the notification. Apple generates one if omitted.
+	// This should be a UUID with 32 lowercase hexadecimal digits.
+	// TODO: use a UUID type.
+	ID string
+
+	// Apple will retry delivery until this time. The default behavior only tries once.
+	Expiration time.Time
+
+	// Allow Apple to group messages together to reduce power consumption.
+	// By default messages are sent immediately.
+	LowPriority bool
+}
+
+// set headers for an HTTP request
+func (h *Headers) set(reqHeader http.Header) {
+	// headers are optional
+	if h == nil {
+		return
+	}
+
+	if h.ID != "" {
+		reqHeader.Set("apns-id", h.ID)
+	} // when omitted, Apple will generate a UUID for you
+
+	if !h.Expiration.IsZero() {
+		reqHeader.Set("apns-expiration", strconv.FormatInt(h.Expiration.Unix(), 10))
+	}
+
+	if h.LowPriority {
+		reqHeader.Set("apns-priority", "5")
+	} // when omitted, the default priority is 10
+}
