+++ push/service_test.go
@@ -9,8 +9,28 @@ import (
 	"testing"
 	"time"
 
+	"github.com/RobotsAndPencils/buford/certificate"
 	"github.com/RobotsAndPencils/buford/push"
 )
+
+func TestNewService(t *testing.T) {
+	const name = "../testdata/cert.p12"
+
+	cert, err := certificate.Load(name, "")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	service, err := push.NewService(push.Development, cert)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	const expectedTopic = ""
+	if service.Topic != expectedTopic {
+		t.Errorf("Expected topic %q, got %q.", expectedTopic, service.Topic)
+	}
+}
 
 func TestPush(t *testing.T) {
 	deviceToken := "c2732227a1d8021cfaf781d71fb2f908c61f5861079a00954a5453f1d0281433"
